+++ resources/views/products/index.blade.php
@@ -1,0 +1,3 @@
+<x-layouts::storefront :title="__('Shop')">
+    <livewire:storefront.product-listing-page />
+</x-layouts::storefront>