+++ resources/views/products/show.blade.php
@@ -1,0 +1,3 @@
+<x-layouts::storefront :title="__('Product')">
+    <livewire:storefront.product-detail-page :product-slug="$product" />
+</x-layouts::storefront>